requirements.txt
flask
elasticsearch
pandas
kafka-python
haxor
